refactor: split ls config, add server config
