+++ client/src/game.ts
@@ -18,15 +18,19 @@ import { Toast } from "@client/ui/toast";
 import { pickRandom } from "@client/utils";
 import { RegModal } from "@client/modal/regModal";
 import { GameMenuModal } from "@client/modal/gameMenuModal";
+import { defaultServerConfigObj, ServerConfig } from "@client/config/server";
 
 export const gameLoader = async () => {
+  const serverCfg = new ServerConfig();
   let wsURLPrefix: string;
   // TODO: what if I connect to different host?
   window.location.protocol === "https:"
     ? (wsURLPrefix = "wss")
     : (wsURLPrefix = "ws");
   const client = new Client(
-    new URL(`${wsURLPrefix}://${window.location.hostname}:5889`)
+    new URL(
+      `${wsURLPrefix}://${serverCfg.getValue("host")}:${serverCfg.getValue("port")}`
+    )
   );
 
   const chat = new Chat(
